+++ Sources/TuistLoader/Models+ManifestMappers/SwiftPackageManagerDependencies+ManifestMapper.swift
@@ -15,7 +15,16 @@ extension TuistGraph.SwiftPackageManagerDependencies {
         let productTypes = manifest.productTypes.mapValues { TuistGraph.Product.from(manifest: $0) }
         let baseSettings = try TuistGraph.Settings.from(manifest: manifest.baseSettings, generatorPaths: generatorPaths)
         let targetSettings = manifest.targetSettings.mapValues { TuistGraph.SettingsDictionary.from(manifest: $0) }
+        let projectOptions: [String: TuistGraph.Project.Options] = manifest
+            .projectOptions
+            .mapValues { .from(manifest: $0) }
 
-        return .init(packages, productTypes: productTypes, baseSettings: baseSettings, targetSettings: targetSettings)
+        return .init(
+            packages,
+            productTypes: productTypes,
+            baseSettings: baseSettings,
+            targetSettings: targetSettings,
+            projectOptions: projectOptions
+        )
     }
 }