+++ projects/tuist/fixtures/app_with_spm_dependencies/Tuist/Dependencies.swift
@@ -1,4 +1,5 @@
 import ProjectDescription
+import ProjectDescriptionHelpers
 
 let packages: [Package] = [
     .package(url: "https://github.com/Alamofire/Alamofire", .upToNextMajor(from: "5.6.0")),
@@ -8,16 +9,17 @@ let packages: [Package] = [
     .package(url: "https://github.com/pointfreeco/swift-composable-architecture", .upToNextMinor(from: "0.34.0")),
     .package(url: "https://github.com/iterable/swift-sdk", .upToNextMajor(from: "6.4.0")),
     .package(url: "https://github.com/Trendyol/ios-components", .revision("c9260bfe203a16a278eca5542c98455eece98aa4")),
+    .package(url: "https://github.com/stripe/stripe-ios", .upToNextMajor(from: "22.4.0")),
+    .local(path: "LocalSwiftPackage"),
 ]
 
 let dependencies = Dependencies(
     swiftPackageManager: .init(
         packages,
-        baseSettings: .settings(configurations: [
-            .debug(name: .debug),
-            .release(name: .release),
-            .release(name: "Internal"),
-        ])
+        baseSettings: .targetSettings,
+        projectOptions: [
+            "LocalSwiftPackage": .options(disableSynthesizedResourceAccessors: false),
+        ]
     ),
     platforms: [.iOS]
 )